+++ tray_linux.go
@@ -1,0 +1,7 @@
+//go:build linux || darwin
+
+package main
+
+func initTray(_ *App) error  { return nil }
+func destroyTray(_ *App)     {}
+func updateTrayTooltip(_ *App) {}